Use FriendOverviewView on Friends list
